+++ xenon-common/src/main/java/com/vmware/xenon/services/common/LuceneDocumentIndexService.java
@@ -18,6 +18,7 @@ import static com.vmware.xenon.services.common.LuceneIndexDocumentHelper.GROUP_B
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
+import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -1053,7 +1054,7 @@ public class LuceneDocumentIndexService extends StatelessService {
             options = EnumSet.noneOf(QueryOption.class);
         }
 
-        if (options.contains(QueryOption.EXPAND_CONTENT)) {
+        if (options.contains(QueryOption.EXPAND_CONTENT) || options.contains(QueryOption.EXPAND_BINARY_CONTENT)) {
             rsp.documents = new HashMap<>();
         }
 
@@ -1743,7 +1744,8 @@ public class LuceneDocumentIndexService extends StatelessService {
         ScoreDoc lastDocVisited = null;
         Set<String> fieldsToLoad = this.fieldsToLoadNoExpand;
         if (options.contains(QueryOption.EXPAND_CONTENT)
-                || options.contains(QueryOption.OWNER_SELECTION)) {
+                || options.contains(QueryOption.OWNER_SELECTION)
+                || options.contains(QueryOption.EXPAND_BINARY_CONTENT)) {
             fieldsToLoad = this.fieldsToLoadWithExpand;
         }
 
@@ -1871,7 +1873,16 @@ public class LuceneDocumentIndexService extends StatelessService {
                 }
             }
 
-            if (options.contains(QueryOption.EXPAND_CONTENT) && !rsp.documents.containsKey(link)) {
+            if (options.contains(QueryOption.EXPAND_BINARY_CONTENT) && !rsp.documents.containsKey(link)) {
+                BytesRef binaryData = getBinaryFromLuceneDocument(d);
+                if (binaryData != null) {
+                    ByteBuffer buffer = ByteBuffer.wrap(binaryData.bytes, binaryData.offset,
+                            binaryData.length);
+                    rsp.documents.put(link, buffer);
+                } else {
+                    logWarning("Binary State not found for %s", link);
+                }
+            } else if (options.contains(QueryOption.EXPAND_CONTENT) && !rsp.documents.containsKey(link)) {
                 if (options.contains(QueryOption.EXPAND_BUILTIN_CONTENT_ONLY)) {
                     ServiceDocument stateClone = new ServiceDocument();
                     state.copyTo(stateClone);
@@ -1883,7 +1894,7 @@ public class LuceneDocumentIndexService extends StatelessService {
             }
 
             if (options.contains(QueryOption.SELECT_LINKS)) {
-                state = processQueryResultsForSelectLinks(s, qs, rsp, d, sd.doc, link, state);
+                processQueryResultsForSelectLinks(s, qs, rsp, d, sd.doc, link, state);
             }
 
             uniques.add(link);
@@ -1973,13 +1984,16 @@ public class LuceneDocumentIndexService extends StatelessService {
         return state;
     }
 
+    private BytesRef getBinaryFromLuceneDocument(Document doc) {
+        return doc.getBinaryValue(LUCENE_FIELD_NAME_BINARY_SERIALIZED_STATE);
+    }
+
     private ServiceDocument getStateFromLuceneDocument(Document doc, String link) {
-        BytesRef binaryStateField = doc.getBinaryValue(LUCENE_FIELD_NAME_BINARY_SERIALIZED_STATE);
+        BytesRef binaryStateField = getBinaryFromLuceneDocument(doc);
         if (binaryStateField == null) {
             logWarning("State not found for %s", link);
             return null;
         }
-
         ServiceDocument state = (ServiceDocument) KryoSerializers.deserializeDocument(
                 binaryStateField.bytes,
                 binaryStateField.offset, binaryStateField.length);